+++ nnforge/accuracy_layer.h
@@ -50,7 +50,7 @@ namespace nnforge
 
 		virtual std::string get_string_for_average_data(
 			const layer_configuration_specific& config,
-			const std::vector<float>& data) const;
+			const std::vector<double>& data) const;
 
 		virtual std::vector<std::string> get_parameter_strings() const;
