+++ nnforge/training_task_state.h
@@ -32,7 +32,7 @@ namespace nnforge
 		network_data::ptr data;
 		network_data::ptr momentum_data;
 		network_data::ptr momentum_data2;
-		std::vector<std::pair<backward_propagation::stat, std::map<std::string, std::pair<layer_configuration_specific, nnforge_shared_ptr<std::vector<float> > > > > > history;
+		std::vector<std::pair<backward_propagation::stat, std::map<std::string, std::pair<layer_configuration_specific, nnforge_shared_ptr<std::vector<double> > > > > > history;
 		std::vector<std::string> comments;
 		unsigned int initial_epoch;
